feat: report migration status
Changes the `onProgress` option to receive feedback on how far each
migration has progressed with a percent complete indicator and a
message to show the user.

If an `onProgress` option is passed, migrations will get a bit slower
as we need to calculate the total volume of work before starting a
migration in order to work out the percent complete.

Also removes datastores from runtime dependencies as these should only
ever be passed in as config.

Fixes #32

BREAKING CHANGES:

- The signature of the `onProgress` callback has changed