feat: migrate mfs root to datastore
We've been storing the mfs root in the repo's 'root' datastore since
forever which means it's a block on the filesystem.

It should have been stored in the datastore instead so add migration 11
which moves the key to the datastore.

BREAKING CHANGE: adds a new migration, should go out as a major